lesson1_Task5 sending

+++ src/main/kotlin/lesson_2/lesson2_task2.kt
@@ -1,15 +1,15 @@
 package org.example.lesson_2
 
 fun main() {
-    val countOfWorkers: Byte = 50   //Число работников
-    val countOfInterns: Byte = 30 //Число стажеров
-    val countOfAll: Int = countOfWorkers + countOfInterns
-    val workerSalary: Short = 30000
-    val internSalary: Short = 20000
-    val allWorkersSalary: Int = countOfWorkers * workerSalary //Общая зп работников
-    val allInternsSalary: Int = countOfInterns * internSalary //Общая зп стажеров
-    val allSalary: Int = allWorkersSalary + allInternsSalary  //Общая ЗП сотрудников
-    val averageSalary: Int = allSalary/countOfAll
+    val countOfWorkers: Float = 50f   //Число работников
+    val countOfInterns: Float = 30f //Число стажеров
+    val countOfAll: Float = countOfWorkers + countOfInterns
+    val workerSalary: Float = 30000f
+    val internSalary: Float = 20000f
+    val allWorkersSalary: Float = countOfWorkers * workerSalary //Общая зп работников
+    val allInternsSalary: Float = countOfInterns * internSalary //Общая зп стажеров
+    val allSalary: Float = allWorkersSalary + allInternsSalary  //Общая ЗП сотрудников
+    val averageSalary: Float = allSalary/countOfAll
     println("Расходы на выплату зарплаты постоянным сотрудникам составляют $allWorkersSalary руб.")
     println("Общие расходы по зарплате по выходу стажеров составляют $allSalary руб.")
     println("Средняя зарплата одного сотрудника составляет $averageSalary руб.")
